Time: 0 ms (100.00%) | Memory: 16.5 MB (25.35%) - LeetSync

+++ 396-rotate-function/rotate-function.c
@@ -1,16 +1,16 @@
 int maxRotateFunction(int* nums, int numsSize) {
-    long sum = 0;
-    long R = 0;
+    int sum = 0;
+    int R = 0;
 
     for (int i = 0; i < numsSize; i++) {
         sum += nums[i];
-        R += (long)i * nums[i];
+        R += (int)i * nums[i];
     }
 
-    long maxVal = R;
+    int maxVal = R;
 
     for (int k = 1; k < numsSize; k++) {
-        R = R + sum - (long)numsSize * nums[numsSize - k];
+        R = R + sum - (int)numsSize * nums[numsSize - k];
         if (R > maxVal) {
             maxVal = R;
         }
